cam merge, pid ul era bun, s a modificat si if ul de la shooter din teleop si merge scorarea

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/prepPositions/OuttakeTest.kt
@@ -57,7 +57,7 @@ class OuttakeTest : LinearOpMode() {
                 lastResetTime = currentTime
             }
 
-            robot.shooter.targetRpm = outtakeConfig.targetRPM
+            robot.shooter.targetRpm = outtakeConfig.targetRPM.toDouble()
 
             robot.shooter.update(timeKeep.deltaTime)
